Move `Id` to its own submodule
