+++ crates/control/src/middleware/sessions.rs
@@ -17,8 +17,8 @@ use crate::context::AppContext;
 use crate::controllers::json_api::{Links, PayloadError, ProblemDetails};
 use crate::controllers::sessions::routes as sessions_routes;
 use crate::models::accounts::Account;
+use crate::models::id::Id;
 use crate::models::sessions::IdentityProvider;
-use crate::models::Id;
 use crate::repo::accounts as accounts_repo;
 use crate::repo::credentials as credentials_repo;
 use crate::services::sessions::{SignedToken, Token};
